requirements.txt
pyobjc-core>=9.0
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-LocalAuthentication>=9.0
pyobjc-framework-Quartz>=9.0
